move configk2 package to configk
