+++ jitsi-utils-kotlin/src/main/kotlin/org/jitsi/utils/configk/examples/ExampleConfigSource.kt
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package org.jitsi.utils.configk2.examples
+package org.jitsi.utils.configk.examples
 
-import org.jitsi.utils.configk2.exception.ConfigPropertyNotFoundException
-import org.jitsi.utils.configk2.exception.ConfigurationValueTypeUnsupportedException
-import org.jitsi.utils.configk2.ConfigSource
+import org.jitsi.utils.configk.exception.ConfigPropertyNotFoundException
+import org.jitsi.utils.configk.exception.ConfigurationValueTypeUnsupportedException
+import org.jitsi.utils.configk.ConfigSource
 import java.time.Duration
 import kotlin.reflect.KClass
